requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.1.1
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.0
cachetools==4.2.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cloud-sql-python-connector==0.5.2
cryptography==38.0.3
Django==3.2.16
django-cors-headers==3.10.1
djangorestframework
djangorestframework-simplejwt
frozenlist==1.2.0
google-auth==2.3.3
idna==3.3
mssql-django==1.1.3
multidict==5.2.0
mysqlclient==2.0.3
Pillow==8.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.0.0
pycparser==2.21
PyJWT==2.3.0
pyodbc==4.0.32
python-tds==1.11.0
pytz==2021.3
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
tzdata==2021.5
urllib3==1.26.7
yarl==1.7.2
